+++ sys-libs/glibc/glibc-2.3.4.20050125.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2005 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/sys-libs/glibc/glibc-2.3.4.20050125.ebuild,v 1.41 2005/05/17 22:33:51 vapier Exp $
+# $Header: /var/cvsroot/gentoo-x86/sys-libs/glibc/glibc-2.3.4.20050125.ebuild,v 1.42 2005/06/11 04:37:24 vapier Exp $
 
 KEYWORDS="~amd64 ~mips ~sparc ~x86 -ppc"
 
@@ -333,13 +333,14 @@ toolchain-glibc_src_install() {
 		mkdir -p ${D}/nptl ${D}$(alt_headers)/nptl
 		make install_root=${D}/nptl install-headers PARALLELMFLAGS="${MAKEOPTS}"
 		pushd ${D}/nptl/$(alt_headers) > /dev/null
-			for i in `find . -type f`; do
-				if ! [ -f ${D}$(alt_headers)/$i ] \
-					|| ! cmp -s $i ${D}$(alt_headers)/$i; then
+		for i in `find . -type f`; do
+			if ! [ -f ${D}$(alt_headers)/$i ] \
+			   || ! cmp -s $i ${D}$(alt_headers)/$i; then
 				mkdir -p ${D}$(alt_headers)/nptl/`dirname $i`
 				cp -a $i ${D}$(alt_headers)/nptl/$i
 			fi
 		done
+		popd > /dev/null
 		rm -rf ${D}/nptl
 	fi
