+++ sys-libs/glibc/glibc-2.3.5.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2005 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/sys-libs/glibc/glibc-2.3.5.ebuild,v 1.24 2005/06/07 01:38:58 vapier Exp $
+# $Header: /var/cvsroot/gentoo-x86/sys-libs/glibc/glibc-2.3.5.ebuild,v 1.25 2005/06/11 04:37:24 vapier Exp $
 
 # Here's how the cross-compile logic breaks down ...
 #  CTARGET - machine that will target the binaries
@@ -377,13 +377,14 @@ toolchain-glibc_src_install() {
 		mkdir -p ${D}/nptl ${D}$(alt_headers)/nptl
 		make install_root=${D}/nptl install-headers PARALLELMFLAGS="${MAKEOPTS} -j1"
 		pushd ${D}/nptl/$(alt_headers) > /dev/null
-			for i in `find . -type f`; do
-				if ! [ -f ${D}$(alt_headers)/$i ] \
-					|| ! cmp -s $i ${D}$(alt_headers)/$i; then
+		for i in `find . -type f`; do
+			if ! [ -f ${D}$(alt_headers)/$i ] \
+			   || ! cmp -s $i ${D}$(alt_headers)/$i; then
 				mkdir -p ${D}$(alt_headers)/nptl/`dirname $i`
 				cp -a $i ${D}$(alt_headers)/nptl/$i
 			fi
 		done
+		popd > /dev/null
 		rm -rf ${D}/nptl
 	fi
